 requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-jose==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
